Add 풀이 삭제 기능 추가 (Api)
- 단일 풀이 관련 기능들 post 모듈로 이동
